chore: create a widget to display a list of informations
