+++ lib/account.dart
@@ -1,9 +1,8 @@
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:thundercard/custom_progress_indicator.dart';
-import 'account_editor.dart';
+import 'package:thundercard/widgets/card_info.dart';
 import 'api/firebase_auth.dart';
 
 class Account extends StatefulWidget {
@@ -19,7 +18,6 @@ class _AccountState extends State<Account> {
   @override
   Widget build(BuildContext context) {
     CollectionReference users = FirebaseFirestore.instance.collection('users');
-    CollectionReference cards = FirebaseFirestore.instance.collection('cards');
 
     return Scaffold(
       appBar: AppBar(
@@ -48,53 +46,8 @@ class _AccountState extends State<Account> {
                         child: Center(
                       child: Padding(
                         padding: const EdgeInsets.all(16.0),
-                        child: StreamBuilder<DocumentSnapshot<Object?>>(
-                          stream: cards.doc(user['my_cards'][0]).snapshots(),
-                          builder: (BuildContext context,
-                              AsyncSnapshot<DocumentSnapshot> snapshot) {
-                            if (snapshot.hasError) {
-                              return const Text('Something went wrong');
-                            }
-                            if (snapshot.connectionState ==
-                                ConnectionState.waiting) {
-                              return const CustomProgressIndicator();
-                            }
-                            dynamic data = snapshot.data;
-                            return Column(
-                              children: [
-                                OutlinedButton(
-                                    onPressed: () {
-                                      Navigator.of(context)
-                                          .push(MaterialPageRoute(
-                                        builder: (context) => AccountEditor(
-                                            data: data,
-                                            cardId: user['my_cards'][0]),
-                                      ));
-                                    },
-                                    child: const Text('プロフィールを編集')),
-                                Text('username: ${data?['name']}'),
-                                data?['bio'] != ''
-                                    ? Text('bio: ${data?['bio']}')
-                                    : Container(),
-                                data?['url'] != ''
-                                    ? Text('URL: ${data?['url']}')
-                                    : Container(),
-                                data?['twitter'] != ''
-                                    ? Text('Twitter: ${data?['twitter']}')
-                                    : Container(),
-                                data?['github'] != ''
-                                    ? Text('GitHub: ${data?['github']}')
-                                    : Container(),
-                                data?['company'] != ''
-                                    ? Text('company: ${data?['company']}')
-                                    : Container(),
-                                data?['email'] != ''
-                                    ? Text('email: ${data?['email']}')
-                                    : Container(),
-                              ],
-                            );
-                          },
-                        ),
+                        child: CardInfo(
+                            cardId: user['my_cards'][0], editable: true),
                       ),
                     )),
                   );
@@ -102,7 +55,6 @@ class _AccountState extends State<Account> {
                 return const Center(child: CustomProgressIndicator());
               }),
           ElevatedButton(
-              // onPressed: () => FirebaseAuth.instance.signOut(),
               onPressed: () async {
                 await FirebaseAuth.instance.signOut();
               },
